requirements.txt
dnspython
rich